backend/app/requirements.txt
fastapi
uvicorn
httpx
pydantic
requests
python-dotenv
aiofiles
python-multipart
firebase-admin
SQLAlchemy
psycopg2-binary

